Struct elvwasm::bccontext::BitcodeContext

pub struct BitcodeContext {
    pub request: Request,
}

This structure encapsulates all communication with the Eluvio content fabric. A new BitcodeContext is automatically created during the processing of the http request. During initialization, all context data is acquired from the http request.

Fields
request: Request
Implementations

impl<'a> BitcodeContext

pub fn new(request: Request) -> BitcodeContext

pub fn log_info(&'a self, s: &str) -> CallResult

pub fn log_debug(&'a self, s: &str) -> CallResult

pub fn log_warn(&'a self, s: &str) -> CallResult

pub fn log_error(&'a self, s: &str) -> CallResult

pub fn write_stream(&'a self, stream: &str, src: &'a [u8]) -> CallResult

write_stream writes a u8 slice of specified length to a fabric stream

Arguments

id- a unique identifier (can use BitcodeContext’s request id)

stream- the fabric stream to write to BitcodeContext::new_stream

src- a u8 slice to write

len - length of the slice to write

Returns

utf8 bytes stream containing json { “written” : bytes }

Example

pub fn read_stream(&'a self, stream_to_read: String, sz: usize) -> CallResult

read_stream reads usize bytes from a fabric stream returning a slice of u8

Arguments

stream_to_read- the fabric stream to read from

sz- usize size of bytes

Returns

utf8 bytes stream containing json { “return” : { “read” : byte-count-read }, “result” : “base64 encoded string” } Example

pub fn callback( &'a self, status: usize, content_type: &str, size: usize ) -> CallResult

callback issues a Callback on the fabric setting up an expectation that the output stream contains a specified sized buffer

Arguments

status- the http status of the call

content-type- output buffer contents

size- size of the output contents

Returns

the checksum as hex-encoded string

Example

pub fn make_success(&'a self, msg: &str) -> CallResult

pub fn make_success_json(&'a self, msg: &Value) -> CallResult

pub fn make_error(&'a self, msg: &'static str) -> CallResult

pub fn make_error_with_kind(&'a self, kind: ErrorKinds) -> CallResult

pub fn make_error_with_error<T>( &'a self, kind: ErrorKinds, _err: T ) -> CallResult

pub fn make_success_bytes(&'a self, msg: &[u8], id: &str) -> CallResult

pub fn call(&'a self, ns: &str, op: &str, msg: &[u8]) -> CallResult

pub fn call_external_bitcode( &'a self, function: &str, args: &Value, object_hash: &str, code_part_hash: &str ) -> CallResult

call_external_bitcode - enables the calling of fabric api’s

Arguments

function - the function to call on the external bitcode

args - the argumaents to pass the external function

object_hash - the content object containing the external bitcode part

code_part_hash - the code part for the external bitcode

Example

use elvwasm::ErrorKinds;
extern crate wapc_guest as guest;
use serde_json::{json, Value};
use std::str;
use guest::CallResult;
use elvwasm::ExternalCallResult;

fn do_something<'s>(bcc: &'s elvwasm::BitcodeContext) -> wapc_guest::CallResult {
    let http_p = &bcc.request.params.http;
    let qp = &http_p.query;
    let id = &bcc.request.id;
    let img_hash = &qp.get("img_hash").ok_or(ErrorKinds::Invalid("img_hash not present".to_string()))?[0];
    let img_obj= &qp.get("img_obj").ok_or(ErrorKinds::Invalid("img_hash not present".to_string()))?[0];
    let tar_hash = &qp.get("tar_hash").ok_or(ErrorKinds::Invalid("tar_hash not present".to_string()))?[0];
    bcc.log_info(&format!("img_hash ={img_hash:?} tar_hash = {tar_hash:?}"))?;
    let params = json!({
        "http" : {
            "verb" : "some",
            "headers": {
                "Content-type": [
                    "application/json"
                ]
            },
            "path" : "/image/default/assets/birds.jpg",
            "query" : {
                "height" : ["200"],
            },
        },
    });
    let img = bcc.call_external_bitcode("image", &params, img_obj, img_hash)?;
    let exr:ExternalCallResult = serde_json::from_slice(&img)?;
    let imgbits = base64::decode(&exr.fout)?;
    Ok(imgbits)
}

pub fn close_stream(&'a self, sid: String) -> CallResult

close_stream closes the fabric stream

sid: the sream id (returned from one of the new_file_stream or new_stream) Returns the checksum as hex-encoded string

Example

pub fn new_stream(&'a self) -> CallResult

new_stream creates a new fabric bitcode stream.

Returns

output u8 of format {"stream_id" : id} where id is a string

Example

pub fn new_file_stream(&'a self) -> CallResult

new_file_stream creates a new fabric file output u8 of format where id and path are strings { “stream_id”: id, “file_name”: path }

pub fn q_download_file( &'a mut self, path: &str, hash_or_token: &str ) -> CallResult

q_download_file : downloads the file stored at the fabric file location path for some content

Arguments

path : fabric file location in the content

hash_or_token : hash for the content containing the file

pub fn q_upload_file( &'a mut self, qwt: &str, input_data: &[u8], path: &str, mime: &str ) -> CallResult

q_upload_file : uploads the input data and stores it at the fabric file location as filetype mime

Arguments

qwt : a fabric write token

input_data : a slice of u8 data

path : fabric file location

mime : MIME type to store the data as (eg gif)

pub fn file_stream_size(&'a self, filename: &str) -> CallResult

file_stream_size computes the current size of a fabric file stream given its stream name filename : the name of the file steam. See new_file_stream.

impl<'a> BitcodeContext

pub fn q_create_content( &'a self, qtype: &str, meta: &HashMap<&str, Value> ) -> CallResult

q_create_content creates a new content object locally. The content will have a write token but will not be comitted to the fabric until calls to finalize and commit are made

Arguments

qtype- a hash for the content type. Can also be “builtin” for built in bitcode

meta- a HashMap containing the initial meta data for the object to be set at ‘/’

Returns

utf8 bytes stream containing json

{ "qid" : "idObj", "qwtoken" : "writeToken"}

pub fn q_list_content(&'a self) -> CallResult

q_list_content calculates a content fabric QList for the context’s libid

Returns

slice of u8 parseable to QList e.g.

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let qlist:elvwasm::QList = bcc.q_list_content().try_into()?;
  let res = vec![];

  // do stuff with the qlist
  Ok(res)
}

pub fn q_list_content_for(&'a self, qlibid: &str) -> CallResult

q_list_content_for calculates a content fabric QList for a given libid

Arguments

qlibid- libid to be listed

Returns

slice of u8 parseable to QList e.g.

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.q_list_content_for(&bcc.request.q_info.qlib_id)?;
  let qlist:elvwasm::QList = serde_json::from_str(std::str::from_utf8(&res).unwrap()).unwrap();
  // do stuff with the qlist
  Ok(res)
}

pub fn q_finalize_content(&'a self, qwtoken: &str) -> CallResult

q_finalize_content finalizes a given write token

Arguments

qwtoken - a write token to finalize

Returns

utf8 bytes stream containing json

{ "qid" : "idObj", "qhash" : "newHash"}

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let mr:elvwasm::ModifyResult = bcc.q_modify_content().try_into()?;
  // ... process content
  let res = vec![];
  let hash = bcc.q_finalize_content(&mr.qwtoken)?;
  Ok(res)
}

Example

pub fn q_commit_content(&'a self, qhash: &str) -> CallResult

q_commit_content finalizes a given write token

Arguments

qhash - a finalized hash

Returns

slice of u8 that is empty e.g.

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.q_commit_content("hq__jd7sd655fffg7HrF76mHDolzzwe")?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn q_system_time(&'a self) -> CallResult

pub fn q_modify_content(&'a self) -> CallResult

q_modify_content enables edit on the implicit content of the context

Returns

utf8 bytes stream containing json

{ "qwtoken" : "writeTokenForEdit"}

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let mr:elvwasm::ModifyResult = bcc.q_modify_content().try_into()?;
  let res = vec![];
  Ok(res)
}

Returns

slice of u8

pub fn q_part_list(&'a self, object_id_or_hash: String) -> CallResult

q_part_list returns a list of parts in a given hash

Arguments

String- object id or hash for the objects parts to be listed

Returns

utf8 bytes containing json QPartList

Example

pub fn write_part_to_stream( &'a self, stream_id: String, qphash: String, qihot: String, offset: i64, length: i64 ) -> CallResult

write_part_to_stream writes the content of a part to to a fabric stream

Arguments

stream_id- stream identifier from new_stream or the like

off- offset into the file (0 based)

len- length of part to write

qphash - part hash to write

Returns

utf8 bytes stream containing json WriteResult

Example

pub fn q_create_part_from_stream( &'a self, qwtoken: &str, stream_id: &str ) -> CallResult

q_create_part_from_stream creates a new part in a writeable object from a context stream. The content will be made locally but not published until finalized and committed

Arguments

qwtoken- a write token to write the part

stream_id- the stream to write BitcodeContext::new_stream

Returns

utf8 bytes stream containing json CreatePartResult

Example

pub fn q_file_to_stream( &'a self, stream_id: &str, path: &str, hash_or_token: &str ) -> CallResult

q_file_to_stream writes the given qfile’s content to the provided stream

Arguments

stream_id- string identifier aquired from BitcodeContext::new_stream

path - string conatining the QFile path

hash_or_token - an optional string with a hash to operate on (defaults to the contexts content)

Returns

utf8 bytes stream containing json WriteResult

Example

pub fn file_to_stream(&'a self, filename: &str, stream: &str) -> CallResult

file_to_stream directs a fabric file (filename) to a fabric stream (stream) filename - name of the fabric file (see new_file_stream) stream - name of the stream that receives the file stream (see new_stream)

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let wr:elvwasm::WriteResult = bcc.file_to_stream("myfile", "someStreamId").try_into()?;
  bcc.log_info(&format!("file written {}", wr.written));
  let res = vec![];
  Ok(res)
}

pub fn q_create_file_from_stream( &'a self, stream_id: &str, qwtoken: &str, path: &str, mime: &str, size: i64 ) -> CallResult

q_create_file_from_stream creates a qfile from the cotents of a bitcode stream

Arguments

stream_id- stream identifier from new_stream or the like

qwtoken- write token (will use context’s if “” is provided)

path- qfile path

mime - MIME type of the file

size - size of the file in bytes

Returns

slice of u8 parseable to QPartInfo

pub fn q_create_q_state_store(&'a self) -> CallResult

q_create_q_state_store creates a new state store in the fabric

Returns

utf8 bytes stream containing a string with the state store id e.g.

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.q_create_q_state_store()?;
  let ssID = std::str::from_utf8(&res)?;
  Ok(res)
}

pub fn q_get_versions(&'a self, qid: &str, with_details: bool) -> CallResult

q_get_versions lists the content versions with additional details for the given qid in context’s libid.

Arguments

qid- id of object to get versions for

with_details- whether to retrieve additional info (currently content type hash, qlib ID, and size stats)

Returns

Vec parseable to QRef e.g.

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.q_get_versions("id_someQID", true)?;
  let qVersions:elvwasm::QRef = serde_json::from_str(std::str::from_utf8(&res).unwrap()).unwrap();
  // Do stuff with qVersions
  Ok(res)
}

pub fn q_checksum_part(&'a self, sum_method: &str, qphash: &str) -> CallResult

q_checksum_part calculates a checksum of a given content part.

Arguments

sum_method- checksum method (“MD5” or “SHA256”)

qphash- hash of the content part to checksum

Returns

the checksum as hex-encoded string

pub fn q_checksum_file( &'a self, sum_method: &str, file_path: &str ) -> CallResult

q_checksum_file calculates a checksum of a file in a file bundle

Arguments

sum_method- checksum method (“MD5” or “SHA256”)

file_path- the path of the file in the bundle

Returns

the checksum as hex-encoded string

pub fn sqmd_set_json(&'a self, path: &'a str, val: &Value) -> CallResult

The following sqmd_* based functions all work on the premise that a bitcode context represents a single content in the fabric. As such, each content has meta data that is directly associated. This meta forms a standard tree at the / root level.

sqmd_set_json ets the metadata at path

Arguments

path : path to the meta data

val : serde_json::Value to set

Returns

error only no success return

use serde_json::json;
fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.sqmd_set_json("/some_key", &json!({"foo" : "bar"}))?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn sqmd_merge_json(&'a self, path: &'a str, json_str: &'a str) -> CallResult

sqmd_merge_json merges the metadata at path

Arguments

path : path to the meta data

val : serde_json::Value to merge

Returns

error only no success return

use serde_json::json;

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.sqmd_merge_json("/some_key", r#"{{"foo" : "bar"}}"#)?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn sqmd_delete_json(&'a self, path: &'a str) -> CallResult

sqmd_delete_json deletes the metadata at path

Arguments

path : path to the meta data

Returns

error only no success return

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.sqmd_delete_json("/some_key")?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn sqmd_clear_json(&'a self, path: &'a str) -> CallResult

sqmd_clear_json clears the metadata at path

Arguments

path : path to the meta data

Returns

nothing only error on failure

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.sqmd_clear_json("/some_key")?; // this will blast some_key and all its descendants
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn sqmd_get_json(&'a self, path: &'a str) -> CallResult

sqmd_get_json gets the metadata at path

Arguments

path : path to the meta data

Returns

UTF8 u8 slice containing json

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.sqmd_get_json("/some_key")?;
  let v:serde_json::Value = serde_json::from_slice(&res.clone()).unwrap();
  let mut meta = v.as_object().unwrap();
  Ok(res)
}

pub fn sqmd_get_json_resolve(&'a self, path: &'a str) -> CallResult

sqmd_get_json_resolve gets the metadata at path resolving all links

Arguments

path : path to the meta data

Returns

UTF8 u8 slice containing json

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.sqmd_get_json_resolve("/some_key")?;
  let v:serde_json::Value = serde_json::from_slice(&res.clone()).unwrap();
  let mut meta = v.as_object().unwrap();
  Ok(res)
}

pub fn sqmd_get_json_external( &'a self, qlibid: &str, qhash: &str, path: &str ) -> CallResult

sqmd_get_json_external gets the metadata at path from another content

Arguments

path : path to the meta data

qhash: hash of external content

Returns

UTF8 u8 slice containing json

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.sqmd_get_json_external("libid4556", "hq_bad2a1ac0a2923ad85e1489736701c06320242a9", "/some_key")?;
  let mut meta_str: String = String::from_utf8(res.clone())?;
  Ok(res)
}

pub fn sqmd_query(&'a self, query: &'a str) -> CallResult

sqmd_query queries the meta-data with the given JSONPath query expression.

Arguments

query : JSONPath query expression

Returns

UTF8 u8 slice containing json

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.sqmd_query("$['some'].value[0].description")?;
  let v:serde_json::Value = serde_json::from_slice(&res.clone()).unwrap();
  let mut meta = v.as_object().unwrap();
  Ok(res)
}

pub fn qss_set(&'a self, qssid: &str, key: &str, val: &str) -> CallResult

qss_set sets data into the Q state store

Arguments

qssid- string identifier aquired from BitcodeContext::q_create_q_state_store

key - string

val - string value to store

VecReturns

Nothing error only

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.q_create_q_state_store()?;
  let ssID = std::str::from_utf8(&res)?;
  bcc.qss_set(ssID, "akey", "avalue")?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

pub fn qss_get(&'a self, qssid: &str, key: &str) -> CallResult

qss_get gets data from the Q state store

Arguments

qssid- string identifier aquired from BitcodeContext::q_create_q_state_store

key - string

Returns

Vec containing string value

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let res = bcc.qss_get("sid_648nfjfh5666nmjejh", "akey")?;
  let strVal = std::str::from_utf8(&res)?;
  Ok(res)
}

pub fn qss_delete(&'a self, qssid: &str, key: &str) -> CallResult

qss_get deletes data from the Q state store

Arguments

qssid- string identifier aquired from BitcodeContext::q_create_q_state_store

key - string

Returns

Nothing error only

fn do_something<'s>(bcc: &'s mut elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.qss_delete("sid_648nfjfh5666nmjejh", "akey")?;
  Ok("SUCCESS".to_owned().as_bytes().to_vec())
}

impl<'a> BitcodeContext

pub fn proxy_http(&'a self, v: Option<Value>) -> CallResult

proxy_http proxies an http request in case of CORS issues

Arguments

v : a JSON Value

use serde_json::json;

fn do_something<'s>(bcc: &'s elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let v = json!({
        "request_parameters" : {
        "url": "https://www.googleapis.com/customsearch/v1?key=AIzaSyCppaD53DdPEetzJugaHc2wW57hG0Y5YWE&q=fabric&cx=012842113009817296384:qjezbmwk0cx",
        "method": "GET",
        "headers": {
          "Accept": "application/json",
          "Content-Type": "application/json"
        }
     }
  });
  bcc.proxy_http(Some(v))
}

Returns

slice of u8

Example

pub fn start_bitcode_lro(&'a self, function: &str, args: &Value) -> CallResult

start_bitcode_lro initiates a long running operation on the fabric. Currently the lro implementation constrains the callback to be in the same bitcode module as the initiator.

Arguments

function : &str the function to call in the current bitcode module

args : JSON value containing the arguments to pass the callback

Returns

slice of u8

Example

impl<'a> BitcodeContext

pub fn new_index_builder(&'a mut self, _v: Value) -> CallResult

new_index_builder Creates a new index builder storing the resultant blob in a fabric part and returning the part hash in the body of an http response

Example

pub fn archive_index_to_part(&'a self, dir: &str) -> CallResult

archive_index_to_part Uses the index constructed in the fabric directory to form a compressed tar storing the resultant blob in a fabric part and returning the part hash

Example

pub fn restore_index_from_part( &'a self, content_hash: &str, part_hash: &str ) -> CallResult

restore_index_from_part Extract the part using the supplied part hash restore an archived tantivy index locating the resultant in a directory on the local node.

Arguments

content_hash : &str the content object hash

part_hash : &str the part hash returned from BitcodeContext::archive_index_to_part

Example

pub fn query_parser_parse_query(&'a self, query: &str) -> CallResult

query_parser_parse_query Queries the index using the assoicated index query context

Example

pub fn builder_add_text_field(&'a self, v: Option<Value>) -> CallResult

builder_add_text_field adds a new text field to a Tantivy index

Arguments

v : a JSON Value

use serde_json::json;

fn do_something<'s>(bcc: &'s elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let v = json!({
    "name":   "title",
    "type":   1,
    "stored": true,
  });
  bcc.builder_add_text_field(Some(v))
}

Example

pub fn builder_build(&'a self, v: Option<Value>) -> CallResult

builder_build builds the new Index Arguments None

use serde_json::json;

fn do_something<'s>(bcc: &'s elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  bcc.builder_build(None)
}

Example

pub fn builder_create_index(&'a self, v: Option<Value>) -> CallResult

builder_create_index create an index from an existing dir

Example

pub fn document_create(&'a self, v: Option<Value>) -> CallResult

document_create create a new document for a given Index

Example

pub fn document_add_text(&'a self, v: Option<Value>) -> CallResult

document_add_text add text to a given document

Example

pub fn document_create_index(&'a self, v: Option<Value>) -> CallResult

document_create_index creates an index given a set of documents

Example

pub fn index_create_writer(&'a self, v: Option<Value>) -> CallResult

index_create_writer creates an index writer

Example

pub fn index_add_document(&'a self, v: Option<Value>) -> CallResult

index_add_document adds a document to the writer

Example

pub fn index_writer_commit(&'a self, v: Option<Value>) -> CallResult

index_writer_commit commits the index

Example

pub fn index_reader_builder_create(&'a self, v: Option<Value>) -> CallResult

index_reader_builder_create creates a new reader builder on an index

Example

pub fn index_reader_searcher(&'a self, v: Option<Value>) -> CallResult

index_reader_searcher creates a new query parser for the index

Example

pub fn reader_builder_query_parser_create( &'a self, v: Option<Value> ) -> CallResult

reader_builder_query_parser_create creates a ReaderBuilder from a QueryParser

Example

pub fn query_parser_for_index(&'a self, v: Option<Value>) -> CallResult

query_parser_for_index executes ForIndex on the QueryParser

Arguments

v : a JSON Value

fn do_something<'s>(bcc: &'s elvwasm::BitcodeContext) -> wapc_guest::CallResult {
  let v = serde_json::from_str(r#"{
        "fields" : ["field1", "field2"]
      }
  }"#).unwrap();
  bcc.query_parser_for_index(v)
}

Returns

slice of u8

Example

pub fn query_parser_search(&'a self, v: Option<Value>) -> CallResult

query_parser_search searches the given QueryParser for the term

Example